Add files via upload
The model uses the data for individual FIs, policy parameters, and macro statistics to project FIs' failures, the DI's disbursements and recoveries, and the DI Fund's net balances for 10 years.  The model expresses the distribution of potential net balances by reporting selected quartiles from the distribution.  The .xlsxm is dated November 2022, using the current version of Excel and VBA for that date.